+++ app/representers/stellar_base/balance_representer.rb
@@ -1,0 +1,8 @@
+module StellarBase
+  class BalanceRepresenter < ApplicationRepresenter
+
+    property :asset_code
+    property :amount
+
+  end
+end